+++ src/main/java/insane96mcp/mobspropertiesrandomness/json/utils/MPRRange.java
@@ -10,6 +10,7 @@ import insane96mcp.mobspropertiesrandomness.exception.InvalidJsonException;
 import insane96mcp.mobspropertiesrandomness.json.IMPRObject;
 import insane96mcp.mobspropertiesrandomness.json.utils.modifier.MPRModifiable;
 import insane96mcp.mobspropertiesrandomness.json.utils.modifier.MPRPosModifier;
+import insane96mcp.mobspropertiesrandomness.json.utils.modifier.MPRTimeExistedModifier;
 import insane96mcp.mobspropertiesrandomness.json.utils.modifier.difficulty.MPRDifficultyModifier;
 import insane96mcp.mobspropertiesrandomness.utils.Logger;
 import net.minecraft.entity.MobEntity;
@@ -24,14 +25,14 @@ public class MPRRange extends MPRModifiable implements IMPRObject {
 	private Float min;
 	private Float max;
 
-	public MPRRange(float min, float max, @Nullable MPRDifficultyModifier difficultyModifier, @Nullable MPRPosModifier posModifier, @Nullable Integer round) {
-		super(difficultyModifier, posModifier, round);
+	public MPRRange(float min, float max, @Nullable MPRDifficultyModifier difficultyModifier, @Nullable MPRPosModifier posModifier, @Nullable MPRTimeExistedModifier timeExistedModifier, @Nullable Integer round) {
+		super(difficultyModifier, posModifier, timeExistedModifier, round);
 		this.min = min;
 		this.max = Math.max(min, max);
 	}
 
 	public MPRRange(float min, float max) {
-		this(min, max, null, null, null);
+		this(min, max, null, null, null, null);
 	}
 
 	public MPRRange(float min) {
@@ -62,6 +63,9 @@ public class MPRRange extends MPRModifiable implements IMPRObject {
 		if (this.posModifier != null)
 			min = this.posModifier.applyModifier(world, entity.position(), min);
 
+		if (this.timeExistedModifier != null && !this.timeExistedModifier.affectsMaxOnly)
+			min = this.timeExistedModifier.applyModifier(world, entity, min);
+
 		return min;
 	}
 
@@ -73,6 +77,9 @@ public class MPRRange extends MPRModifiable implements IMPRObject {
 
 		if (this.posModifier != null)
 			max = this.posModifier.applyModifier(world, entity.position(), max);
+
+		if (this.timeExistedModifier != null)
+			max = this.timeExistedModifier.applyModifier(world, entity, max);
 
 		return max;
 	}
@@ -101,7 +108,7 @@ public class MPRRange extends MPRModifiable implements IMPRObject {
 		public MPRRange deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
 			if (json.isJsonPrimitive())
 				return new MPRRange(json.getAsFloat());
-			return new MPRRange(json.getAsJsonObject().get("min").getAsFloat(), json.getAsJsonObject().get("max").getAsFloat(), context.deserialize(json.getAsJsonObject().get("difficulty_modifier"), MPRDifficultyModifier.class), context.deserialize(json.getAsJsonObject().get("pos_modifier"), MPRPosModifier.class), context.deserialize(json.getAsJsonObject().get("round"), Integer.class));
+			return new MPRRange(json.getAsJsonObject().get("min").getAsFloat(), json.getAsJsonObject().get("max").getAsFloat(), context.deserialize(json.getAsJsonObject().get("difficulty_modifier"), MPRDifficultyModifier.class), context.deserialize(json.getAsJsonObject().get("pos_modifier"), MPRPosModifier.class), context.deserialize(json.getAsJsonObject().get("time_existed_modifier"), MPRTimeExistedModifier.class), context.deserialize(json.getAsJsonObject().get("round"), Integer.class));
 		}
 	}
 }