+++ src/main/java/insane96mcp/mobspropertiesrandomness/json/utils/modifier/MPRModifiable.java
@@ -16,22 +16,29 @@ public abstract class MPRModifiable implements IMPRObject {
 	@SerializedName("pos_modifier")
 	public MPRPosModifier posModifier;
 
-	//Rounds the result to this value
+	//Applied after pos modifier
+	@SerializedName("time_existed_modifier")
+	public MPRTimeExistedModifier timeExistedModifier;
+
+	//Rounds the final result to this value
 	public Integer round;
 
-	public MPRModifiable(MPRDifficultyModifier difficultyModifier, MPRPosModifier posModifier, Integer round) {
+	public MPRModifiable(MPRDifficultyModifier difficultyModifier, MPRPosModifier posModifier, MPRTimeExistedModifier timeExistedModifier, Integer round) {
 		this.difficultyModifier = difficultyModifier;
 		this.posModifier = posModifier;
+		this.timeExistedModifier = timeExistedModifier;
 		this.round = round;
 	}
 
 	public void validate(final File file) throws InvalidJsonException {
-		if (difficultyModifier != null)
-			difficultyModifier.validate(file);
+		if (this.difficultyModifier != null)
+			this.difficultyModifier.validate(file);
 
-		if (posModifier != null)
-			posModifier.validate(file);
+		if (this.posModifier != null)
+			this.posModifier.validate(file);
 
+		if (this.timeExistedModifier != null)
+			this.timeExistedModifier.validate(file);
 	}
 
 	public float round(float value) {
@@ -42,6 +49,6 @@ public abstract class MPRModifiable implements IMPRObject {
 
 	@Override
 	public String toString() {
-		return String.format("difficulty_modifier: %s, pos_modifier: %s, round: %s", difficultyModifier, posModifier, round);
+		return String.format("difficulty_modifier: %s, pos_modifier: %s, time_existed_modifier: %s, round: %s", difficultyModifier, posModifier, timeExistedModifier, round);
 	}
 }