+++ changelog.md
@@ -1,9 +1,11 @@
 # Changelog
 
 ## Upcoming
+* Added `conditions` tag to mobs and presets. `spawner_behaviour` and `structure_behaviour` have been moved into `conditions`  
+  Currently contains `is_baby` which checks if the mob is baby and `nbt` to check for raw nbt of the mob.  
+  This also fixes presets counting as applied even if spawner or structure behaviour prevented the preset from applying.
 * NBT for items is now parsed on load.  
-  This should make the applying of items faster due to not having to parse the NBT each time an item is set
-* 
+  This should make the applying of items faster due to not having to parse the NBT each time an item is set 
 
 ## 3.1.3
 * Mobs with EXCLUSIVE presets now correctly apply normal properties if no preset is chosen